requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

langchain>=0.3.0
langchain-core>=0.3.0
langchain-classic>=1.0.0        

langchain-chroma>=0.2.0          
langchain-google-genai>=2.0.0    
langchain-huggingface>=0.1.2     
chromadb>=0.6.0

pypdf>=5.0.0                     

python-dotenv>=1.0.0             
